+++ src/App.css
@@ -1,128 +1,42 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-:root {
-  --color-blue: #4635b1;
-  --color-purple: #b771e5;
-  --color-green: #aeea94;
-  --color-yellow: #fffbca;
-}
-
-main {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  background: var(--color-purple);
-  height: 100vh;
-  width: 100vw;
-  gap: 5px;
-}
-
-.card {
-  background-color: var(--color-yellow);
-  display: grid;
-  place-items: center;
-}
-
-button {
-  background-color: var(--color-green);
-  color: var(--color-blue);
-}
-
-span {
-  background-color: var(--color-purple);
-  color: var(--color-yellow);
-  padding: 4px 8px;
-  border-radius: 12px;
-  cursor: default;
-}
-.tooltip-container {
-  position: relative;
-}
-.tooltip-text {
-  color: black;
-  border: 1px solid black;
-  font-size: 12px;
-  width: 120px;
-  text-align: center;
-}
-.Top {
-  position: absolute;
-  top: -50px;
-}
-.Right {
-  position: absolute;
-  left: 130px;
-  top: 2px;
-  right: -120px;
-  background-color: #fffbca;
-  z-index: 1;
-}
-.Left {
-  position: absolute;
-  right: 130px;
-  top: 2px;
-  background-color: #fffbca;
-  z-index: 1;
-}
-.Bottom {
-  position: absolute;
-  bottom: -50px;
-}
-
-.sidebar-container {
-  position: relative;
-  height: 100vh;
-  width: 100vw;
-  overflow: hidden;
-}
-
-.content {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  gap: 5px;
-  height: 100%;
-  width: 100%;
-  background: var(--color-purple);
+#root {
+    max-width: 1280px;
+    margin: 0 auto;
+    padding: 2rem;
+    text-align: center;
+  }
   
-}
-
-.content.dimmed {
-  opacity: 0.5;
-  pointer-events: none; /
-}
-
-.sidebar {
-  position: absolute;
-  top: 0;
-  right: -300px; 
-  width: 300px;
-  height: 100%;
-  background: #f0f0f0;
-  box-shadow: -2px 0 5px rgba(0, 0, 0, 0.2);
-  padding: 20px;
-  z-index: 1000; 
-  transition: right 0.3s ease-in-out; 
-}
-
-.visible {
-  right: 0; 
-}
-
-.close-btn {
-  background: #d9534f;
-  color: white;
-  border: none;
-  padding: 10px 15px;
-  cursor: pointer;
-  margin-bottom: 20px;
-}
-
-.close-btn:hover {
-  background: #c9302c;
-}
-
+  .logo {
+    height: 6em;
+    padding: 1.5em;
+    will-change: filter;
+    transition: filter 300ms;
+  }
+  .logo:hover {
+    filter: drop-shadow(0 0 2em #646cffaa);
+  }
+  .logo.react:hover {
+    filter: drop-shadow(0 0 2em #61dafbaa);
+  }
+  
+  @keyframes logo-spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+  
+  @media (prefers-reduced-motion: no-preference) {
+    a:nth-of-type(2) .logo {
+      animation: logo-spin infinite 20s linear;
+    }
+  }
+  
+  .card {
+    padding: 2em;
+  }
+  
+  .read-the-docs {
+    color: #888;
+  }
